+++ tembo-operator/src/cloudnativepg/mod.rs
@@ -3,8 +3,8 @@ pub mod clusters;
 pub(crate) mod cnpg;
 // pub(crate) mod cnpg_backups;
 pub mod cnpg_utils;
+pub mod hibernate;
 pub(crate) mod placement;
 pub mod poolers;
 mod scheduledbackups;
-
 pub const VOLUME_SNAPSHOT_CLASS_NAME: &str = "cnpg-snapshot-class";
